refactor: moves files for better reading
